Adds some test translation keys
